requirements.txt
selenium
Pillow
numpy
requests
